do not serialize a null embedded property.
Right now when an embedded value is null HATEOAS serializes it anyway.
Thinking it's a wrong layout I decided to fix it. Like that a null
embedded value is not serialized.

I also take into account the shouldSerializeNull from the context of
JMS.

[see: X]

take into account the context given to the serialiser